Add a cli tool for calc
