+++ calc/__init__.py
@@ -1,3 +1,4 @@
 from .app import repl_executor, script_executor
 from .executor import calc_execute
 from .metadata import APPNAME, VERSION
+from .cli import run_cli
