+++ main.py
@@ -1,8 +1,6 @@
-from calc import repl_executor, script_executor
-from calc.utils import print_help_msg, print_version
 import colorama
 import os
-import sys
+from calc import run_cli
 
 
 def warning():
@@ -14,32 +12,8 @@ def warning():
     print()
 
 
-def arg_parse():
-    if len(sys.argv) >= 1:
-        return sys.argv[1:]
-    return False
-
-
-def main():
-    args = arg_parse()
-    if args:
-        if args[0].startswith("--"):
-            if args[0] == "--version":
-                print_version()
-                return 0
-            elif args[0] == "--help":
-                print_help_msg()
-                return 0
-            else:
-                raise ValueError(f"Invalid argument \"{args[0]}\"")
-        else:
-            return script_executor(args[0])
-    else:
-        return repl_executor()
-
-
 if __name__ == "__main__":
     env = os.environ['ENVIRONMENT']
     if env == "Dev":
         warning()
-    main()
+    run_cli()
