+++ pyproject.toml
@@ -8,8 +8,14 @@ authors = ["Jeel Dobariya <dobariyaj34@gmail.com>"]
 python = ">=3.10.0,<3.12"
 colorama = "^0.4.6"
 
+[tool.poetry.group.lint]
+optional = true
+
 [tool.poetry.group.lint.dependencies]
 flake8 = "^7.0.0"
+
+[tool.poetry.scripts]
+calc = 'calc:run_cli'
 
 [tool.pyright]
 # https://github.com/microsoft/pyright/blob/main/docs/configuration.md
